+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/fir/AnvilInteropTest.kt
@@ -13,6 +13,7 @@ import dev.zacsweers.metro.compiler.createGraphWithNoArgs
 import dev.zacsweers.metro.compiler.generatedMetroGraphClass
 import kotlin.test.Test
 import org.jetbrains.kotlin.name.ClassId
+import org.junit.Ignore
 
 class AnvilInteropTest : MetroCompilerTest() {
 
@@ -123,6 +124,91 @@ class AnvilInteropTest : MetroCompilerTest() {
       val contributedInterface = graph.callProperty<Any>("contributedInterface")
       assertThat(contributedInterface).isNotNull()
       assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl3")
+    }
+  }
+
+  // Covers a bug where the codebase has a Dagger-Anvil-ContributesBinding usage that outranks a
+  // Metro-ContributesBinding usage with an explicit type and multiple supertypes. Basically ensures
+  // that rank processing doesn't rely on the outranked binding using the Dagger-Anvil annotation.
+  @Test
+  fun `ranked binding processing supports outranked bindings using Metro's @ContributesBinding`() {
+    compile(
+      source(
+        """
+          interface ContributedInterface
+
+          interface OtherInterface
+
+          // Having two supertypes and an explicit binding type with Metro's @ContributesBinding
+          // annotation is the key piece of this repro
+          @ContributesBinding(AppScope::class, binding = binding<ContributedInterface>())
+          object Impl1 : ContributedInterface, OtherInterface
+
+          @com.squareup.anvil.annotations.ContributesBinding(AppScope::class, rank = 10)
+          object Impl2 : ContributedInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+        """
+          .trimIndent()
+      ),
+      options = metroOptions.withAnvilContributesBinding(),
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl2")
+    }
+  }
+
+  @Ignore(
+    "Can enable once type arguments are saved to metadata - https://youtrack.jetbrains.com/issue/KT-76954/Some-type-arguments-are-not-saved-to-metadata-in-FIR"
+  )
+  @Test
+  fun `ranked binding processing supports outranked bindings using Metro's @ContributesBinding from downstream module`() {
+    val libCompilation =
+      compile(
+        source(
+          """
+          interface ContributedInterface
+
+          interface OtherInterface
+
+          // Having two supertypes and an explicit binding type with Metro's @ContributesBinding
+          // annotation is the key piece of this repro
+          @ContributesBinding(AppScope::class, binding<ContributedInterface>())
+          @Inject
+          class Impl1 : ContributedInterface, OtherInterface
+        """
+            .trimIndent()
+        )
+      )
+
+    compile(
+      source(
+        """
+          @com.squareup.anvil.annotations.ContributesBinding(AppScope::class, rank = 10)
+          @Inject
+          class Impl2 : ContributedInterface
+
+          @DependencyGraph(scope = AppScope::class)
+          interface ExampleGraph {
+            val contributedInterface: ContributedInterface
+          }
+
+        """
+          .trimIndent()
+      ),
+      previousCompilationResult = libCompilation,
+      options = metroOptions.withAnvilContributesBinding(),
+    ) {
+      val graph = ExampleGraph.generatedMetroGraphClass().createGraphWithNoArgs()
+      val contributedInterface = graph.callProperty<Any>("contributedInterface")
+      assertThat(contributedInterface).isNotNull()
+      assertThat(contributedInterface.javaClass.name).isEqualTo("test.Impl2")
     }
   }